+++ src/navigation/HomeStack.tsx
@@ -1,75 +1,26 @@
-import * as React from 'react';
-import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
-import { MaterialCommunityIcons } from '@expo/vector-icons';
+import React from 'react';
+
+import { createStackNavigator } from '@react-navigation/stack';
 import HomeScreen from 'screens/Home/HomeScreen';
-import UploadScreen from 'screens/Documents/Documents';
-import FormsScreen from 'screens/Forms/FormsScreen';
-import ScheduleScreen from 'screens/Schedule/ScheduleScreen';
-import { Colors } from 'assets/Colors';
+import TestScreen from 'screens/Home/TestScreen';
 
-const Tab = createMaterialBottomTabNavigator();
+const Stack = createStackNavigator();
 
-const BottomTabs = () => {
+const HomeStack = () => {
   return (
-    <Tab.Navigator
-      initialRouteName="Home"
-      barStyle={{ backgroundColor: Colors.brandBlue }}
-      activeColor={Colors.navbarPink}
-    >
-      <Tab.Screen
-        name="Home"
+    <Stack.Navigator>
+      <Stack.Screen
+        name="HomeScreen"
+        options={{ headerShown: false }}
         component={HomeScreen}
-        options={{
-          tabBarIcon: ({ color }) => (
-            <MaterialCommunityIcons
-              name="home-outline"
-              color={color}
-              size={26}
-            />
-          ),
-        }}
       />
-      <Tab.Screen
-        name="Forms"
-        component={FormsScreen}
-        options={{
-          tabBarIcon: ({ color }) => (
-            <MaterialCommunityIcons
-              name="note-text-outline"
-              color={color}
-              size={26}
-            />
-          ),
-        }}
+      <Stack.Screen
+        name="TestScreen"
+        options={{ headerShown: false }}
+        component={TestScreen}
       />
-      <Tab.Screen
-        name="Upload"
-        component={UploadScreen}
-        options={{
-          tabBarIcon: ({ color }) => (
-            <MaterialCommunityIcons
-              name="file-upload-outline"
-              color={color}
-              size={26}
-            />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name="Schedule"
-        component={ScheduleScreen}
-        options={{
-          tabBarAccessibilityLabel: 'Schedule',
-          tabBarIcon: ({ color }) => (
-            <MaterialCommunityIcons
-              name="calendar-month-outline"
-              color={color}
-              size={26}
-            />
-          ),
-        }}
-      />
-    </Tab.Navigator>
+    </Stack.Navigator>
   );
 };
-export default BottomTabs;
+
+export default HomeStack;
