+++ src/navigation/RootNavigator.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { StyleSheet, ActivityIndicator, View } from 'react-native';
-import HomeStack from 'navigation/HomeStack';
+import HomeStack from 'navigation/TabStack';
 import firebase from 'database/clientApp';
 import AuthStack from './AuthStack';
