+++ src/screens/Home/TestScreen.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { ScrollPageContainer, InnerPageContainer } from 'screens/styles';
+
+const TestScreen = ({ navigation }: any) => {
+  return (
+    <ScrollPageContainer>
+      <InnerPageContainer>
+        <p>Hello</p>
+      </InnerPageContainer>
+    </ScrollPageContainer>
+  );
+};
+
+export default TestScreen;
